plugins: scan: introduce scan plugin
